requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==25.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0